+++ spec/grape/router/route_spec.rb
@@ -22,6 +22,42 @@ RSpec.describe Grape::Router::Route do
     it { is_expected.to be_a(Grape::Router::BaseRoute) }
   end
 
+  describe 'route attributes' do
+    subject(:route) do
+      described_class.new(endpoint, :get, pattern, options, forward_match: false,
+                                                            version: 'v1', namespace: '/things', prefix: '/api',
+                                                            requirements: { id: /\d+/ }, anchor: false, settings: { a: 1 })
+    end
+
+    it 'exposes them as readers' do
+      expect(route.version).to eq('v1')
+      expect(route.namespace).to eq('/things')
+      expect(route.prefix).to eq('/api')
+      expect(route.requirements).to eq(id: /\d+/)
+      expect(route.anchor).to be(false)
+      expect(route.settings).to eq(a: 1)
+    end
+
+    it 'does not leak them into the options Hash' do
+      %i[version namespace prefix requirements anchor settings].each do |key|
+        expect(route.options).not_to have_key(key)
+      end
+    end
+
+    context 'when omitted' do
+      subject(:route) { described_class.new(endpoint, :get, pattern, options, forward_match: false) }
+
+      it 'defaults to nil' do
+        expect(route.version).to be_nil
+        expect(route.namespace).to be_nil
+        expect(route.prefix).to be_nil
+        expect(route.requirements).to be_nil
+        expect(route.anchor).to be_nil
+        expect(route.settings).to be_nil
+      end
+    end
+  end
+
   describe '#match?' do
     subject { instance.match?(input) }